[edn/rtl] Handle backpressure from CSRNG without an output FIFO
Upstream OpenTitan commit 75df0aeca193d16b08bcf92fbd20f2e188bad1ee

This commit enables the EDN to handle backpressure from CSRNG by introducing
handshakes and gating the propagation of words of the commands. Words now only
move down the pipeline when CSRNG is ready to accept the next word.

To enable this change, this commit introduces some tweaks to the main SM.
The main SM now has some new states in which the EDN is notified that
the corresponding boot commands can be sent.

This commit also handles the fields of the sw_cmd_sts accordingly.

The following files are changed in OpenTitan commit but not included:
- hw/ip/edn/rtl/edn_core.sv
- hw/ip/edn/rtl/edn_main_sm.sv

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>