[util/reggen] reverse order of substruct generation
Upstream OpenTitan commit 61a237e1973d6a29d377c2be871139553c7d344f

Change the order of the sub-structs in the reg_pkg.
The least significant sub-structs used to be at the most
significant position. Now the least significant sub-structs
should be at the least significant position. This allows
for direct assignment of the whole struct without the order
of the sub-structs being reversed.

In the various IP hjson files there are sometimes registers
with multiple fields. If there are multiple fields, they are
listed in order of significance (LSB to MSB). Our python
scripts parse the fields in the same order as in the hjson.
util/reggen/reg_pkg.sv.tpl then creates the reg_pkg based on
those fields. For each register a type definition is made in
the form of a struct. If multiple fields are contained in
a single register, they are represented by substructs.
Those substructs however, are generated in the same order
as in the hjson. The issue here is that for SV the LSB of a
struct are listed last in a struct. This means that casting a
value directly to a type defined in the reg_pkg reverses the
order of the fields. To allow for direct casting without
reversing the order of the fields this change generates the
substructs as intended by the hjson.

The following files are changed in OpenTitan commit but not included:
- util/reggen/reg_pkg.sv.tpl
- hw/ip/adc_ctrl/rtl/adc_ctrl_reg_pkg.sv
- hw/ip/aon_timer/rtl/aon_timer_reg_pkg.sv
- hw/ip/clkmgr/rtl/clkmgr_reg_pkg.sv
- hw/ip/edn/rtl/edn_reg_pkg.sv
- hw/ip/flash_ctrl/rtl/flash_ctrl_reg_pkg.sv
- hw/ip/gpio/rtl/gpio_reg_pkg.sv
- hw/ip/hmac/rtl/hmac_reg_pkg.sv
- hw/ip/i2c/rtl/i2c_reg_pkg.sv
- hw/ip/keymgr/rtl/keymgr_reg_pkg.sv
- hw/ip/lc_ctrl/rtl/lc_ctrl_reg_pkg.sv
- hw/ip/otbn/rtl/otbn_reg_pkg.sv
- hw/ip/otp_ctrl/rtl/otp_ctrl_reg_pkg.sv
- hw/ip/pattgen/rtl/pattgen_reg_pkg.sv
- hw/ip/pinmux/rtl/pinmux_reg_pkg.sv
- hw/ip/pwm/rtl/pwm_reg_pkg.sv
- hw/ip/rstmgr/rtl/rstmgr_reg_pkg.sv
- hw/ip/rv_core_ibex/rtl/rv_core_ibex_reg_pkg.sv
- hw/ip/rv_timer/rtl/rv_timer_reg_pkg.sv
- hw/ip/spi_device/rtl/spi_device_reg_pkg.sv
- hw/ip/spi_host/rtl/spi_host_reg_pkg.sv
- hw/ip/sram_ctrl/rtl/sram_ctrl_reg_pkg.sv
- hw/ip/sysrst_ctrl/rtl/sysrst_ctrl_reg_pkg.sv
- hw/ip/trial1/rtl/trial1_reg_pkg.sv
- hw/ip/uart/rtl/uart_reg_pkg.sv
- hw/ip/usbdev/rtl/usbdev_reg_pkg.sv
- hw/top_earlgrey/ip/sensor_ctrl/rtl/sensor_ctrl_reg_pkg.sv
- hw/top_earlgrey/ip_autogen/alert_handler/rtl/alert_handler_reg_pkg.sv
- hw/top_earlgrey/ip_autogen/pwrmgr/rtl/pwrmgr_reg_pkg.sv
- hw/top_earlgrey/ip/clkmgr/rtl/autogen/clkmgr_reg_pkg.sv
- hw/top_earlgrey/ip/flash_ctrl/rtl/autogen/flash_ctrl_reg_pkg.sv
- hw/top_earlgrey/ip/pinmux/rtl/autogen/pinmux_reg_pkg.sv
- hw/top_earlgrey/ip/rstmgr/rtl/autogen/rstmgr_reg_pkg.sv
- hw/ip/kmac/rtl/kmac_reg_pkg.sv

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>