[entropy_src/doc] Document behavior of SHA3 conditioner engine
Upstream OpenTitan commit 809ba9c2b1c62305fe6bd62b97b26c171da8722f

Previously, the documentation and comments in the RTL suggested that
the number of bits the conditioner takes to produce a single 384-bit
seed would be fixed to 2048 bits. However, this is not true. In a
nutshell, the number of bits consumed is equal to the configured health
test window size. This commit clarifies this in the documentation as
well as in the RTL.

In addition, this commit also documents in the RTL when the SHA3
conditioner block is actually active. So far, this the assumption was
the conditioner was only active if explicitly triggered by the main
state machine upon finishing checking the next health test window.
However, the SHA3 engine is also active whenever 832 bits have been
received. This condition is not tracked by the main state machine which
is the reason for the CS AES Halt interface currently being broken.

See also lowRISC/OpenTitan#17941.

The following files are changed in OpenTitan commit but not included:
- hw/ip/entropy_src/README.md
- hw/ip/entropy_src/doc/registers.md
- hw/ip/entropy_src/doc/theory_of_operation.md

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>