[kmac, sha3] Add REQ/ACK interface to delay Keccak operations
Upstream OpenTitan commit 15f75fa3436c90551056a528e20e279905cf46c0

This interface is required to be able to properly delay operations of
the SHA3-based conditioner inside the ENTROPY_SRC block until CSRNG
confirms its internal AES core is halted.

Previously, it has been tried to handle this as part of the main FSM
of ENTROPY_SRC outside of the SHA3 core. However, this is not sufficient
as in the case of ENTROPY_SRC, most SHA3 operations are actually
triggered by the sha3pad instance inside the SHA3 core.

This is related to lowRISC/OpenTitan#17941.

The following files are changed in OpenTitan commit but not included:
- hw/ip/kmac/rtl/kmac.sv
- hw/ip/kmac/rtl/kmac_reduced.sv

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>