[entropy_src] Fix CS AES Halt interface
Upstream OpenTitan commit 21adde16af04c2d1f46b307b36854686f39084e8

Now that the SHA3 core features an interface to directly implement the
functionality to correctly sync with the AES core inside CSRNG in order
to avoid both of them becoming active at the same time, most of the
custom logic for this feature inside ENTROPY_SRC can be removed.

This resolves lowRISC/OpenTitan#17941.

The following files are changed in OpenTitan commit but not included:
- hw/ip/entropy_src/doc/interfaces.md
- hw/ip/entropy_src/doc/programmers_guide.md
- hw/top_earlgrey/data/autogen/top_earlgrey.gen.hjson
- sw/device/lib/dif/dif_entropy_src.h
- hw/ip/entropy_src/dv/env/seq_lib/entropy_src_rng_vseq.sv

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>