[entropy_src] Align enable delay module with fixed CS AES Halt interface
Upstream OpenTitan commit 2b8870ccd4666e9f40defe1668fe24bde915d7ba

Previously, the CS AES Halt interface was only active when the SHA3
engine was performing the final Process operation for which the main SM
always acknowledges the completion with a done pulse. After fixing
the interface to always be active when the SHA3 engine is actively
processing data, the main SM only sends the done pulse for a minority
of SHA3 operations. This can cause the enable delay module to block
the re-enablement of the entropy pipeline after disabling as it keeps
waiting for a done pulse that is never going to arrive.

This commit fixes this issue by using the sha3_block_processed signal
instead of the done pulse. This signal is sent by the SHA3 engine
whenever the processing of a block finishes.

The following files are changed in OpenTitan commit but not included:
- hw/ip/entropy_src/dv/cov/entropy_src_cov_if.sv

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>