[entropy_src] Add FIFO to deal with backpressure from the conditioner
Upstream OpenTitan commit 74a6e9b8b0c80408b02386fd73ac7efae711d374

This commit adds a 32-bit wide distribution FIFO of configurable depth.
The FIFO is added between the postht FIFO, the observe FIFO, the bypass
FIFO and the precon FIFO. Its main purpose is to buffer entropy bits
while the conditioner is busy such that we don't have to drop entropy
bits from the hardware pipeline.

Dropping entropy bits is not a big issue per se as it's allowed by the
spec (when done after the health tests and in a way such that number
of samples going into the conditioner is fixed). Also, under normal
operating conditions, noise source samples arrive at very low rate and
dropping bits should not be needed.

However, verifying that the `correct` entropy bits are dropped is hard
and seems impossible for our current DV environment as it requires to
very accurately model the hardware pipeline which is undesirable. Thus,
the safest approach is to add this new distribution FIFO and tune its
depth parameter to handle potential backpressure from the conditioner
such that dropping bits is not necessary.

The following files are changed in OpenTitan commit but not included:
- hw/ip/entropy_src/doc/registers.md
- hw/top_earlgrey/data/autogen/top_earlgrey.gen.hjson
- hw/top_earlgrey/rtl/autogen/top_earlgrey.sv
- sw/device/lib/dif/dif_entropy_src_unittest.cc
- hw/ip/entropy_src/dv/cov/entropy_src_cov_if.sv
- hw/ip/entropy_src/dv/env/entropy_src_env_cfg.sv
- hw/ip/entropy_src/dv/env/entropy_src_env_pkg.sv
- hw/ip/entropy_src/dv/sva/entropy_src_assert_if.sv
- hw/ip/entropy_src/dv/env/seq_lib/entropy_src_err_vseq.sv

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>