[csrng/rtl] Change and assign cmd_sts signal
Upstream OpenTitan commit 042e0b915865a8473d4d53a4976720f70e799c58

This commit changes the sw_cmd_sts and hw_cmd_sts signals to multiple bits.
0 represents a successful acknowledgement, whereas any value other than 0
represents an error status.
The sts signal is now also able to anounce to the EDN that the CSRNG main
SM is hanging due to an invalid acmd.
This commit also aligns the EDN rtl with the new sts response.
Furthermore a new csrng interface pkg is created to avoid circular dependencies.

This commit also aligns the EDN and CSRNG DV, documentation and SW with
the new STS signal.

The following files are changed in OpenTitan commit but not included:
- hw/dv/sv/csrng_agent/csrng_agent_cfg.sv
- hw/dv/sv/csrng_agent/csrng_device_driver.sv
- hw/dv/sv/csrng_agent/csrng_if.sv
- hw/ip/csrng/doc/registers.md
- hw/ip/edn/data/edn.hjson
- hw/ip/edn/doc/registers.md
- hw/ip/edn/rtl/edn_core.sv
- hw/ip/edn/rtl/edn_reg_pkg.sv
- hw/ip/edn/rtl/edn_reg_top.sv
- sw/device/lib/dif/dif_csrng.c
- sw/device/lib/dif/dif_csrng.h
- sw/device/lib/dif/dif_csrng_unittest.cc
- sw/device/lib/dif/dif_edn.c
- sw/device/lib/dif/dif_edn_unittest.cc
- sw/device/lib/testing/csrng_testutils.c
- sw/device/lib/testing/entropy_testutils.c
- hw/ip/edn/dv/cov/edn_cov_if.sv
- hw/ip/edn/dv/env/edn_env_cfg.sv
- hw/ip/edn/dv/env/edn_env_pkg.sv
- hw/ip/edn/dv/env/edn_scoreboard.sv
- hw/ip/edn/dv/tests/edn_base_test.sv
- sw/device/lib/crypto/drivers/entropy.c
- hw/ip/edn/dv/env/seq_lib/edn_alert_vseq.sv

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>