[csrng/rtl] Add command out of sequence cmd_sts error
Upstream OpenTitan commit 903b31917104b6d9e22f0eb41941a72415109a8a

This commit adds a command out of sequence error to the cmd_sts signal
going to the consumers as a command response. The cmd_sts signal will
now signal an error when a command is issued out of sequence. This can
mean one of two things. Either an instantiate command is issued on an
already instantiated state or any command other than an instantiate is
issued before the state was instantiated.

This commit also adds the documentation and DV alignment for
the rtl changes.

The following files are changed in OpenTitan commit but not included:
- hw/ip/csrng/csrng.core
- sw/device/tests/power_virus_systemtest.c
- hw/ip/csrng/doc/registers.md
- hw/ip/edn/data/edn.hjson
- hw/ip/edn/doc/registers.md
- hw/ip/edn/rtl/edn_reg_pkg.sv
- hw/ip/edn/rtl/edn_reg_top.sv
- sw/device/lib/dif/dif_csrng.h
- hw/ip/edn/dv/env/edn_env_pkg.sv
- hw/ip/csrng/dv/env/seq_lib/csrng_alert_vseq.sv
- hw/ip/csrng/dv/env/seq_lib/csrng_base_vseq.sv
- hw/ip/csrng/dv/env/seq_lib/csrng_intr_vseq.sv
- hw/ip/csrng/dv/env/seq_lib/csrng_smoke_vseq.sv

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>