[entropy_src] Reduce depth of Observe FIFO
Upstream OpenTitan commit 54661e6ec7dcc8493421310b8b4822a9c31a38fd

Halving the depth of the observe FIFO allows saving 1024 FFs. With a
FIFO size of 1 kbit, it is still possible to buffer a full, contiguous
1024-sample window in single-channel mode. For the multi-channel mode,
an experiment on the engineering sample showed that Ibex is fast enough
in reading out the FIFO to avoid it overflowing during read out.

The following files are changed in OpenTitan commit but not included:
- sw/device/tests/entropy_src_ast_rng_req_test.c
- sw/device/tests/entropy_src_fw_observe_many_contiguous.c
- sw/device/tests/entropy_src_fw_override_test.c
- sw/device/tests/entropy_src_fw_ovr_test.c
- hw/ip/entropy_src/doc/entsrc_blk_diag.svg
- hw/ip/entropy_src/doc/programmers_guide.md
- hw/ip/entropy_src/doc/registers.md
- sw/device/tests/sim_dv/ast_clk_rst_inputs.c

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>