[csrng] Add SVAs to check for unexpected pushes to the genbits FIFO
Upstream OpenTitan commit 0c397d707354adca6cf168bc857d22e6781af390

The ready signal of the genbits FIFO is not considered when pushing the
FIFO which may raise concerns. However, by inspecting the command stage
FSM, one can see that the FSM only ever triggers the generation of more
bits if the genbits FIFO is empty.

This commit documents these findings in the RTL and adds some SVAs to
check that 1) the FSM indeed only ever enters the stage where it
requests more bits after finding an empty FIFO and 2) that the genbits
FIFO is only ever pushed in the in the FSM states where this can be
expected.

This resolves lowRISC/OpenTitan#16563.

The following files are changed in OpenTitan commit but not included:
- hw/ip/csrng/dv/env/seq_lib/csrng_err_vseq.sv

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>