[csrng/rtl] Move cmd checks to the cmd stage
Upstream OpenTitan commit b9e0388ab1932d781cf51702437cfa55dae15716

Having the command checks in the main SM doesn't make a lot of sense.
This causes a lot of issues including a hang condition inside the cmd stage
in case of generate commands with cmd len greater than 1. For this reason
it is best to not even start processing commands that violate our checks.

The following files are changed in OpenTitan commit but not included:
- hw/ip/csrng/doc/registers.md
- hw/ip/csrng/doc/theory_of_operation.md
- sw/device/lib/dif/dif_csrng.h
- hw/ip/csrng/dv/env/csrng_env_pkg.sv
- hw/ip/csrng/dv/env/seq_lib/csrng_alert_vseq.sv

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>