[csrng] Make reseed counters readable at any point in time
Upstream OpenTitan commit c903d65d25cd3eb982f0eb5772110fd0bcdd053d

Previously, the reseed counters were only readable if the
otp_en_csrng_sw_app_read fuses were enabled and if the
CTRL.READ_INT_STATE field was enabled, i.e., if the full internal state
was readable.

However, the reseed counter should be readable even if the internal
state is not readable because this allows firmware to track the age
of the different CSRNG contexts. The reseed counter itself is not
sensitive information as it doesn't allow inferring the values of
Key and V.

The following files are changed in OpenTitan commit but not included:
- hw/ip/csrng/doc/registers.md
- hw/ip/csrng/dv/env/csrng_scoreboard.sv

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>