[csrng] Flush command FIFO after signaling ACK errors
Upstream OpenTitan commit 6381071bac44a3c82bb602fe0dfd0063b57a3953

The decision whether a command is invalid is taken based on the command
header but EDN might continue and send the additional data as well (it
might already be in the FIFO). Previously, CSRNG would just pop the
invalid command header from the FIFO and leave any additional data in
the FIFO which can then erroneously be interpreted as the next command
header.

Instead, we have to absorb the full invalid command before we can
continue. This commit modifies the design to keep emptying the FIFO
until the command interface is idle.

This is related to lowRISC/OpenTitan#23526.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>