[entropy_src] Add SVA to ensure FIFOs are empty when triggering SHA
Upstream OpenTitan commit 57841895813aa65b7360785361979352f25111f5

If any of the postht, distr or precon FIFOs are non-empty when
triggering the conditioner, there is risk of leaving entropy bits tested
as part of the current window inside the pipeline instead of absorbing
them into the current seed. This would be problematic.

This is related to lowRISC/OpenTitan#21855.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Gary Guo <gary.guo@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>